move update method to profile resource and update the view
